backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
sqlalchemy>=2.0.23
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
httpx>=0.25.0
aiosqlite>=0.19.0
greenlet>=3.0.0
